backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
supabase
PyPDF2
requests
pydantic